+++ scripts/apps/player-spell-book.mjs
@@ -1,11 +1,11 @@
-import { CANTRIP_RULES, FLAGS, MODULE, SETTINGS, TEMPLATES } from '../constants.mjs';
-import * as actorSpellUtils from '../helpers/actor-spells.mjs';
+import { FLAGS, MODULE, SETTINGS, TEMPLATES } from '../constants.mjs';
 import * as filterUtils from '../helpers/filters.mjs';
 import * as formElements from '../helpers/form-elements.mjs';
 import * as genericUtils from '../helpers/generic-utils.mjs';
-import * as discoveryUtils from '../helpers/spell-discovery.mjs';
-import * as formattingUtils from '../helpers/spell-formatting.mjs';
 import { SpellManager } from '../helpers/spell-preparation.mjs';
+import { SpellbookState } from '../helpers/state/spellbook-state.mjs';
+import { SpellbookFilterHelper } from '../helpers/ui/spellbook-filters.mjs';
+import { SpellbookUI } from '../helpers/ui/spellbook-ui.mjs';
 import { WizardSpellbookManager } from '../helpers/wizard-spellbook.mjs';
 import { log } from '../logger.mjs';
 import { PlayerFilterConfiguration } from './player-filter-configuration.mjs';
@@ -18,11 +18,6 @@ const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
  * Player-facing spell book application for managing prepared spells
  */
 export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
-  /* -------------------------------------------- */
-  /*  Static Properties                           */
-  /* -------------------------------------------- */
-
-  /** @override */
   static DEFAULT_OPTIONS = {
     id: `player-${MODULE.ID}`,
     tag: 'form',
@@ -55,221 +50,77 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
     }
   };
 
-  /** @override */
   static PARTS = {
-    container: {
-      template: TEMPLATES.PLAYER.CONTAINER
-    },
-    sidebar: {
-      template: TEMPLATES.PLAYER.SIDEBAR
-    },
-    navigation: {
-      template: TEMPLATES.PLAYER.TAB_NAV
-    },
-    spellsTab: {
-      template: TEMPLATES.PLAYER.TAB_SPELLS,
-      scrollable: ['']
-    },
-    wizardTab: {
-      template: TEMPLATES.PLAYER.TAB_WIZARD_SPELLBOOK,
-      scrollable: ['']
-    },
-    footer: {
-      template: TEMPLATES.PLAYER.FOOTER
-    }
+    container: { template: TEMPLATES.PLAYER.CONTAINER },
+    sidebar: { template: TEMPLATES.PLAYER.SIDEBAR },
+    navigation: { template: TEMPLATES.PLAYER.TAB_NAV },
+    spellsTab: { template: TEMPLATES.PLAYER.TAB_SPELLS, scrollable: [''] },
+    wizardTab: { template: TEMPLATES.PLAYER.TAB_WIZARD_SPELLBOOK, scrollable: [''] },
+    footer: { template: TEMPLATES.PLAYER.FOOTER }
   };
-
-  /* -------------------------------------------- */
-  /*  Instance Properties                         */
-  /* -------------------------------------------- */
-
-  /** The actor this spell book is for */
-  actor = null;
-
-  /** Loading state */
-  isLoading = true;
-
-  /** Spell levels data */
-  spellLevels = [];
-
-  /** Class name for spellcasting */
-  className = '';
-
-  /** Spell preparation statistics */
-  spellPreparation = { current: 0, maximum: 0 };
-
-  /** Cantrip manager instance */
-  spellManager = null;
-
-  /** Tracking state for cantrip swapping during level-up */
-  _cantripTracking = {
-    originalChecked: new Set(),
-    hasUnlearned: false,
-    hasLearned: false,
-    unlearned: null,
-    learned: null
-  };
-
-  /** Current cantrip count from UI state */
-  _uiCantripCount = 0;
-
-  _newlyCheckedCantrips = new Set();
-
-  /** Wizard spellbook manager instance */
-  wizardManager = null;
-
-  wizardSpellbookCache = null;
 
   get title() {
     return game.i18n.format('SPELLBOOK.Application.ActorTitle', { name: this.actor.name });
   }
 
-  /* -------------------------------------------- */
-  /*  Constructor                                 */
-  /* -------------------------------------------- */
-
   constructor(actor, options = {}) {
     super(options);
-    log(3, `Initializing PlayerSpellBook for ${actor.name}`);
     this.actor = actor;
     this.spellManager = new SpellManager(actor);
-    this.wizardManager = null;
-    this._newlyCheckedCantrips = new Set();
-    this._isLongRest = false;
+    this.wizardManager = genericUtils.isWizard(actor) ? new WizardSpellbookManager(actor) : null;
 
-    // Flag to ensure we only initialize the wizard spellbook once
+    // Initialize helpers
+    this._stateManager = new SpellbookState(this);
+    this.ui = new SpellbookUI(this);
+    this.filterHelper = new SpellbookFilterHelper(this);
+
+    // Set up state properties
+    this.isLoading = true;
+    this.spellLevels = [];
+    this.className = '';
+    this.spellPreparation = { current: 0, maximum: 0 };
+    this._newlyCheckedCantrips = new Set();
+    this._isLongRest = this.actor.getFlag(MODULE.ID, FLAGS.WIZARD_LONG_REST_TRACKING) || false;
     this._wizardInitialized = false;
 
-    const localizedWizardName = game.i18n.localize('SPELLBOOK.Classes.Wizard').toLowerCase();
-    const wizardClass = actor.items.find((i) => i.type === 'class' && i.name.toLowerCase() === localizedWizardName);
-    const forceWizardMode = actor.getFlag(MODULE.ID, FLAGS.FORCE_WIZARD_MODE);
-
-    if (wizardClass || forceWizardMode) {
-      this.wizardManager = new WizardSpellbookManager(actor);
-      log(3, `Initialized wizard manager for ${actor.name} - ${wizardClass ? 'Has wizard class' : 'Force wizard mode enabled'}`);
-    }
-
-    // Check for long rest flag - persisted between sessions
-    this._isLongRest = actor.getFlag(MODULE.ID, FLAGS.WIZARD_LONG_REST_TRACKING) || false;
-    log(3, `Long rest flag for ${actor.name}: ${this._isLongRest ? 'ACTIVE' : 'NOT SET'}`);
-    log(3, `All actor flags: ${JSON.stringify(actor.flags[MODULE.ID])}`);
-
-    if (this._isLongRest) {
-      log(3, `Long rest cantrip swap mode active for ${actor.name}`);
-    }
-
     // Initialize tab groups
-    if (!this.tabGroups['spellbook-tabs']) {
-      this.tabGroups['spellbook-tabs'] = 'spellstab';
-    }
+    if (!this.tabGroups['spellbook-tabs']) this.tabGroups['spellbook-tabs'] = 'spellstab';
 
-    // Listen for flag changes to refresh the display
+    // Register flag change hook
     this._flagChangeHook = Hooks.on('updateActor', (updatedActor, changes) => {
       if (updatedActor.id !== this.actor.id) return;
-
-      // Check for spell book related flag changes
       if (changes.flags?.[MODULE.ID]) {
         const changedFlags = Object.keys(changes.flags[MODULE.ID]);
         const cantripFlagChanged = changedFlags.some((flag) => [FLAGS.CANTRIP_RULES, FLAGS.ENFORCEMENT_BEHAVIOR, FLAGS.FORCE_WIZARD_MODE].includes(flag));
         const wizardFlagChanged = changedFlags.some((flag) => [FLAGS.WIZARD_SPELLBOOK, FLAGS.WIZARD_LEARNED_SPELLS, FLAGS.WIZARD_COPIED_SPELLS].includes(flag));
 
         if ((cantripFlagChanged || wizardFlagChanged) && this.rendered) {
-          log(3, 'Wizard or cantrip flags changed, re-rendering spell book');
           this.render(false);
         }
       }
     });
   }
 
-  /* -------------------------------------------- */
-  /*  Core Application Methods                    */
-  /* -------------------------------------------- */
-
-  /** @override */
   async _prepareContext(options) {
-    log(3, 'Preparing PlayerSpellBook context');
     const context = this._createBaseContext();
     if (this.isLoading) return context;
     context.spellLevels = this.spellLevels.map((level) => {
       const processedLevel = { ...level };
-      processedLevel.spells = level.spells.map((spell) => {
-        const processedSpell = foundry.utils.deepClone(spell);
-        if (!spell.compendiumUuid) {
-          log(2, 'No uuid for:', { spell: spell, coreId: spell.flags?.core?.sourceId, dndId: spell.flags?.dnd5e?.sourceId, itemId: spell.system?.parent?._source._stats.compendiumSource });
-          spell.compendiumUuid = spell.flags?.core?.sourceId || spell.flags?.dnd5e?.sourceId || spell.system?.parent?._source._stats.compendiumSource || '';
-        }
-        const ariaLabel =
-          spell.preparation.prepared ? game.i18n.format('SPELLBOOK.Preparation.Unprepare', { name: spell.name }) : game.i18n.format('SPELLBOOK.Preparation.Prepare', { name: spell.name });
-        const checkbox = formElements.createCheckbox({
-          name: `spellPreparation.${spell.compendiumUuid}`,
-          checked: spell.preparation.prepared,
-          disabled: spell.preparation.disabled,
-          ariaLabel: ariaLabel
-        });
-        checkbox.id = `prep-${spell.compendiumUuid}`;
-        checkbox.dataset.uuid = spell.compendiumUuid;
-        checkbox.dataset.name = spell.name;
-        checkbox.dataset.ritual = spell.filterData?.isRitual || false;
-        checkbox.dataset.wasPrepared = spell.preparation.prepared;
-        if (spell.preparation.disabled && spell.preparation.disabledReason) {
-          checkbox.dataset.tooltip = game.i18n.localize(spell.preparation.disabledReason);
-        }
-        processedSpell.preparationCheckboxHtml = formElements.elementToHtml(checkbox);
-        if (this.wizardManager && this.wizardManager.isWizard) {
-          processedSpell.inWizardSpellbook = this.wizardSpellbookCache?.includes(spell.compendiumUuid) || false;
-        }
-        return processedSpell;
-      });
+      processedLevel.spells = level.spells.map((spell) => this._processSpellForDisplay(spell));
       return processedLevel;
     });
-    context.isWizard = false;
-    if (this.wizardManager && this.wizardManager.isWizard) {
-      context.isWizard = true;
-      context.wizardSpellbookCount = this.wizardSpellbookCache?.length || 0;
-      context.wizardRulesVersion = this.spellManager.getSettings().rules;
-      if (this._tabData && this._tabData.wizardtab) {
-        context.wizardTotalSpellbookCount = this._tabData.wizardtab.wizardTotalSpellbookCount || 0;
-        context.wizardFreeSpellbookCount = this._tabData.wizardtab.wizardFreeSpellbookCount || 0;
-        context.wizardRemainingFreeSpells = this._tabData.wizardtab.wizardRemainingFreeSpells || 0;
-        context.wizardHasFreeSpells = this._tabData.wizardtab.wizardHasFreeSpells || false;
-      }
-    }
     context.activeTab = this.tabGroups['spellbook-tabs'];
     context.tabs = this._getTabs();
+    context.isWizard = !!this.wizardManager?.isWizard;
     if (context.isWizard) {
-      context.tabs.wizardtab = {
-        id: 'wizardtab',
-        label: game.i18n.format('SPELLBOOK.Tabs.WizardSpells', { class: this.className }),
-        group: 'spellbook-tabs',
-        cssClass: this.tabGroups['spellbook-tabs'] === 'wizardtab' ? 'active' : '',
-        icon: 'fa-solid fa-book-spells'
-      };
+      this._addWizardContextData(context);
     }
-    context.hasMultipleTabs = false;
-    if (context.tabs) {
-      const tabCount = Object.keys(context.tabs).length;
-      context.hasMultipleTabs = tabCount > 1;
-    }
+    context.hasMultipleTabs = Object.keys(context.tabs).length > 1;
     context.filters = this._prepareFilters();
     return context;
   }
 
   _createBaseContext() {
-    // Get filter configuration
-    let filterConfig = game.settings.get(MODULE.ID, SETTINGS.FILTER_CONFIGURATION);
-    if (!Array.isArray(filterConfig) || !filterConfig.length) {
-      filterConfig = DEFAULT_FILTER_CONFIG;
-    }
-
-    // Create empty filters structure for loading state
-    const emptyFilters = {
-      search: null,
-      dropdowns: [],
-      checkboxes: [],
-      range: null
-    };
-
-    // Create buttons array with GM-only cantrip config
     const buttons = [
       {
         type: 'submit',
@@ -293,7 +144,7 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
       isLoading: this.isLoading,
       spellLevels: this.spellLevels || [],
       className: this.className || '',
-      filters: this.isLoading ? emptyFilters : this._getFilterState(),
+      filters: this.isLoading ? { search: null, dropdowns: [], checkboxes: [], range: null } : this.filterHelper.getFilterState(),
       spellSchools: CONFIG.DND5E.spellSchools,
       buttons: buttons,
       actorId: this.actor.id,
@@ -302,106 +153,141 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
     };
   }
 
-  /** @override */
+  _processSpellForDisplay(spell) {
+    const processedSpell = foundry.utils.deepClone(spell);
+
+    if (!spell.compendiumUuid) {
+      spell.compendiumUuid = genericUtils.getSpellUuid(spell);
+    }
+
+    const ariaLabel = spell.preparation.prepared ? game.i18n.format('SPELLBOOK.Preparation.Unprepare', { name: spell.name }) : game.i18n.format('SPELLBOOK.Preparation.Prepare', { name: spell.name });
+
+    const checkbox = formElements.createCheckbox({
+      name: `spellPreparation.${spell.compendiumUuid}`,
+      checked: spell.preparation.prepared,
+      disabled: spell.preparation.disabled,
+      ariaLabel: ariaLabel
+    });
+
+    checkbox.id = `prep-${spell.compendiumUuid}`;
+    checkbox.dataset.uuid = spell.compendiumUuid;
+    checkbox.dataset.name = spell.name;
+    checkbox.dataset.ritual = spell.filterData?.isRitual || false;
+    checkbox.dataset.wasPrepared = spell.preparation.prepared;
+
+    if (spell.preparation.disabled && spell.preparation.disabledReason) {
+      checkbox.dataset.tooltip = game.i18n.localize(spell.preparation.disabledReason);
+    }
+
+    processedSpell.preparationCheckboxHtml = formElements.elementToHtml(checkbox);
+
+    if (this.wizardManager?.isWizard) {
+      processedSpell.inWizardSpellbook = this._stateManager.wizardSpellbookCache?.includes(spell.compendiumUuid) || false;
+    }
+
+    return processedSpell;
+  }
+
+  _addWizardContextData(context) {
+    context.wizardSpellbookCount = this._stateManager.wizardSpellbookCache?.length || 0;
+    context.wizardRulesVersion = this.spellManager.getSettings().rules;
+
+    // Add wizard tab
+    context.tabs.wizardtab = {
+      id: 'wizardtab',
+      label: game.i18n.format('SPELLBOOK.Tabs.WizardSpells', { class: this.className }),
+      group: 'spellbook-tabs',
+      cssClass: this.tabGroups['spellbook-tabs'] === 'wizardtab' ? 'active' : '',
+      icon: 'fa-solid fa-book-spells'
+    };
+
+    // Add tab data if available
+    if (this._stateManager.tabData?.wizardtab) {
+      context.wizardTotalSpellbookCount = this._stateManager.tabData.wizardtab.wizardTotalSpellbookCount || 0;
+      context.wizardFreeSpellbookCount = this._stateManager.tabData.wizardtab.wizardFreeSpellbookCount || 0;
+      context.wizardRemainingFreeSpells = this._stateManager.tabData.wizardtab.wizardRemainingFreeSpells || 0;
+      context.wizardHasFreeSpells = this._stateManager.tabData.wizardtab.wizardHasFreeSpells || false;
+    }
+  }
+
   async _onRender(context, options) {
     super._onRender?.(context, options);
+    this._setupContentWrapper();
 
-    if (!this.element.querySelector('.content-wrapper')) {
-      // Select all direct children except the tabs
-      const tabsNav = this.element.querySelector('.window-content > nav.tabs.tabs-right');
-
-      // Create wrapper and move elements inside
-      const wrapper = document.createElement('div');
-      wrapper.className = 'content-wrapper';
-
-      // Find elements to wrap
-      const elementsToWrap = [this.element.querySelector('.sidebar'), this.element.querySelector('.spell-book-container'), this.element.querySelector('.window-content > footer')].filter((el) => el); // Filter out any null elements
-
-      // Insert wrapper before the first element to move
-      if (elementsToWrap.length && elementsToWrap[0].parentNode) {
-        elementsToWrap[0].parentNode.insertBefore(wrapper, elementsToWrap[0]);
-
-        // Move elements into wrapper
-        elementsToWrap.forEach((el) => wrapper.appendChild(el));
-
-        // Ensure tabs stay outside the wrapper
-        if (tabsNav && tabsNav.parentNode === wrapper) {
-          this.element.querySelector('.window-content').appendChild(tabsNav);
-        }
-      }
-    }
     try {
-      // Set sidebar state based on user preference
-      this._setSidebarState();
+      this.ui.setSidebarState();
 
       if (this.isLoading) {
-        log(3, 'Spell book still loading, setting up loading UI');
         this.element.classList.add('loading');
-        this._disableInputsWhileLoading();
-        this._positionFooter();
-
-        // Start loading data
+        this.ui.disableInputsWhileLoading();
+        this.ui.positionFooter();
         await this._loadSpellData();
-        return; // Return early to avoid additional setup
+        return;
       } else {
         this.element.classList.remove('loading');
       }
 
-      // Setup wizard-specific elements if needed
-      if (this.wizardManager?.isWizard) {
-        // Initialize the journal only once
-        if (!this._wizardInitialized) {
-          this._wizardInitialized = true;
-          await this.wizardManager.getOrCreateSpellbookJournal().catch((err) => {
-            log(1, `Error initializing wizard spellbook journal: ${err.message}`);
-          });
-        }
+      // Initialize wizard spellbook if needed
+      if (this.wizardManager?.isWizard && !this._wizardInitialized) {
+        this._wizardInitialized = true;
+        await this.wizardManager.getOrCreateSpellbookJournal().catch((err) => {
+          log(1, `Error initializing wizard spellbook journal:`, err);
+        });
       }
 
-      // Set up UI elements
-      this._positionFooter();
-      this._setupFilterListeners();
-      this._setupPreparationListeners();
-      this._applyCollapsedLevels();
+      // Set up UI
+      this.ui.positionFooter();
+      this.ui.setupFilterListeners();
+      this.ui.setupPreparationListeners();
+      this.ui.applyCollapsedLevels();
       this._applyFilters();
-      this._updateSpellPreparationTracking();
-      this._setupCantripUI();
-      this._updateSpellCounts();
+      this.ui.updateSpellPreparationTracking();
+      this.ui.setupCantripUI();
+      this.ui.updateSpellCounts();
     } catch (error) {
       log(1, 'Error in _onRender:', error);
     }
   }
 
-  /** @override */
+  _setupContentWrapper() {
+    if (!this.element.querySelector('.content-wrapper')) {
+      const tabsNav = this.element.querySelector('.window-content > nav.tabs.tabs-right');
+      const wrapper = document.createElement('div');
+      wrapper.className = 'content-wrapper';
+
+      const elementsToWrap = [this.element.querySelector('.sidebar'), this.element.querySelector('.spell-book-container'), this.element.querySelector('.window-content > footer')].filter((el) => el);
+
+      if (elementsToWrap.length && elementsToWrap[0].parentNode) {
+        elementsToWrap[0].parentNode.insertBefore(wrapper, elementsToWrap[0]);
+        elementsToWrap.forEach((el) => wrapper.appendChild(el));
+
+        if (tabsNav && tabsNav.parentNode === wrapper) {
+          this.element.querySelector('.window-content').appendChild(tabsNav);
+        }
+      }
+    }
+  }
+
   _onClose() {
     try {
-      // Clean up any long rest flags
       if (this._isLongRest) {
         this.actor.unsetFlag(MODULE.ID, FLAGS.WIZARD_LONG_REST_TRACKING);
       }
 
-      // Remove the flag change hook
       if (this._flagChangeHook) {
         Hooks.off('updateActor', this._flagChangeHook);
       }
 
-      log(3, `Closed PlayerSpellBook for ${this.actor.name}`);
       super._onClose?.();
     } catch (error) {
       log(1, 'Error in _onClose:', error);
     }
   }
 
-  /**
-   * Generate tabs for the spell book application
-   * @returns {object} Tab configuration
-   * @private
-   */
   _getTabs() {
     const tabGroup = 'spellbook-tabs';
 
-    // Define tabs - starting with spells tab
-    const tabs = {
+    return {
       spellstab: {
         id: 'spellstab',
         label: game.i18n.format('SPELLBOOK.Tabs.Spells', { name: this.actor.name }),
@@ -410,37 +296,23 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
         icon: 'fa-solid fa-book-open'
       }
     };
-
-    return tabs;
   }
 
-  /** @override */
   changeTab(tabName, groupName, options = {}) {
-    // First, let the parent method handle the basic tab switching
     super.changeTab(tabName, groupName, options);
 
-    // If switching to spell tab and we need to reload (because we learned a new spell)
     if (tabName === 'spellstab' && this._spellsTabNeedsReload && this.wizardManager?.isWizard) {
-      log(3, `Reloading spells tab data after learning new spells`);
-
-      // Reset flag
       this._spellsTabNeedsReload = false;
 
-      // If we have pre-processed tab data, reload the prep tab from cached wizard data
-      if (this._tabData) {
-        // Store the current UI state we want to preserve
+      if (this._stateManager.tabData) {
         const collapsedLevels = Array.from(this.element.querySelectorAll('.spell-level.collapsed')).map((el) => el.dataset.level);
 
-        // Reload the spell data for the prep tab
         this._loadSpellData().then(() => {
-          // After loading, restore collapsed levels
           setTimeout(() => {
             collapsedLevels.forEach((levelId) => {
               const levelEl = this.element.querySelector(`.spell-level[data-level="${levelId}"]`);
               if (levelEl) {
                 levelEl.classList.add('collapsed');
-
-                // Update ARIA state
                 const heading = levelEl.querySelector('.spell-level-heading');
                 if (heading) heading.setAttribute('aria-expanded', 'false');
               }
@@ -451,1175 +323,49 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
       }
     }
 
-    // If we're a wizard and have pre-processed tab data, update the content
-    if (this.wizardManager?.isWizard && this._tabData && this._tabData[tabName]) {
-      log(3, `Switching to wizard tab: ${tabName}`);
-
-      // Update the spell data for the selected tab
-      this.spellLevels = this._tabData[tabName].spellLevels;
-      this.spellPreparation = this._tabData[tabName].spellPreparation;
-
-      // Re-render the application to display the new data
+    if (this.wizardManager?.isWizard && this._stateManager.tabData && this._stateManager.tabData[tabName]) {
+      this.spellLevels = this._stateManager.tabData[tabName].spellLevels;
+      this.spellPreparation = this._stateManager.tabData[tabName].spellPreparation;
       this.render(false);
     }
   }
 
-  /** @override */
   _configureRenderOptions(options) {
-    // Get default parts configuration
     super._configureRenderOptions(options);
 
-    // Start with the basic UI parts
     options.parts = ['container', 'sidebar', 'navigation', 'spellsTab', 'footer'];
 
-    // Add wizard tab only if the actor is a wizard
     if (this.wizardManager?.isWizard) {
       options.parts.push('wizardTab');
     }
   }
 
-  /* -------------------------------------------- */
-  /*  Data Loading Methods                        */
-  /* -------------------------------------------- */
-
   async _loadSpellData() {
     try {
-      log(3, `Loading spell data for ${this.actor.name}`);
+      await this._stateManager.initialize();
+      this.isLoading = this._stateManager.isLoading;
+      this.spellLevels = this._stateManager.spellLevels;
+      this.className = this._stateManager.className;
+      this.spellPreparation = this._stateManager.spellPreparation;
 
-      // Initialize cantrip manager and flags
-      await this.spellManager.initializeFlags();
-
-      // Cache wizard spellbook data if needed
-      if (this.wizardManager?.isWizard) {
-        await this._cacheWizardSpellbook();
-      }
-
-      // Load spellcasting class
-      const classItem = await this._loadSpellcastingClass();
-      if (!classItem) {
-        log(1, 'No spellcasting class found for actor');
-        this.isLoading = false;
-        this.render(false);
-        return;
-      }
-
-      // Check for cantrip level-up
-      const cantripLevelUp = this.spellManager.checkForLevelUp();
-      if (cantripLevelUp) {
-        const settings = this.spellManager.getSettings();
-        const message = settings.rules === CANTRIP_RULES.DEFAULT ? 'SPELLBOOK.Cantrips.LevelUpDefault' : 'SPELLBOOK.Cantrips.LevelUpModern';
-
-        ui.notifications.info(game.i18n.localize(message));
-        log(3, `Cantrip level-up detected, using rules: ${settings.rules}`);
-      }
-
-      // For wizards, we need to load two different spell sources
-      if (this.wizardManager?.isWizard) {
-        await this._loadWizardSpellData(classItem);
-      } else {
-        // Non-wizard characters - just load the class spell list
-        await this._loadRegularSpellData(classItem);
-      }
-
-      log(3, `Completed loading spell data for ${this.actor.name}`);
+      this.render(false);
     } catch (error) {
       log(1, 'Error loading spell data:', error);
-    } finally {
       this.isLoading = false;
       this.render(false);
     }
   }
 
-  async _loadWizardSpellData(classItem) {
-    try {
-      log(3, `Loading wizard-specific spell data for ${this.actor.name}`);
-      const className = classItem.name.toLowerCase();
-      const classUuid = classItem.uuid;
-      const actorLevel = this.actor.system.details.level;
-      const maxSpellLevel = discoveryUtils.calculateMaxSpellLevel(actorLevel, classItem.system.spellcasting);
-
-      // 1. Get the full class spell list from compendium source
-      log(3, `Loading full wizard class spell list`);
-      const fullSpellList = await discoveryUtils.getClassSpellList(className, classUuid, null); // Pass null for actor to avoid getting the personal spellbook
-      if (!fullSpellList || !fullSpellList.size) {
-        log(1, 'No spells found in class spell list');
-        return;
-      }
-      log(3, `Found ${fullSpellList.size} spells in full class spell list`);
-
-      // 2. Get the wizard's personal spellbook from their journal
-      log(3, `Loading wizard's personal spellbook`);
-      const personalSpellbook = await this.wizardManager.getSpellbookSpells();
-      log(3, `Found ${personalSpellbook.length} spells in wizard's personal spellbook`);
-
-      // Store the full spell list for filtering in the process step
-      this._fullWizardSpellList = new Set(fullSpellList);
-
-      // 3. Load all spells from both sources
-      const allUuids = new Set([...fullSpellList, ...personalSpellbook]);
-      log(3, `Loading ${allUuids.size} total unique spells`);
-
-      // Load up to max level for learned spells, but allow seeing 1st level spells for new characters
-      const effectiveMaxLevel = Math.max(1, maxSpellLevel);
-      const spellItems = await actorSpellUtils.fetchSpellDocuments(allUuids, effectiveMaxLevel);
-
-      if (!spellItems || !spellItems.length) {
-        log(1, 'No spell items could be loaded');
-        return;
-      }
-      log(3, `Loaded ${spellItems.length} spell items up to level ${effectiveMaxLevel}`);
-
-      // 4. Process the spells for both tabs
-      await this._processWizardSpells(spellItems, classItem, personalSpellbook);
-    } catch (error) {
-      log(1, 'Error loading wizard spell data:', error);
-    }
-  }
-
-  async _loadRegularSpellData(classItem) {
-    try {
-      const className = classItem.name.toLowerCase();
-      const classUuid = classItem.uuid;
-
-      // Load spell list as normal
-      const spellList = await discoveryUtils.getClassSpellList(className, classUuid, this.actor);
-      if (!spellList || !spellList.size) {
-        log(1, 'No spells found in spell list');
-        return;
-      }
-      log(3, `Found ${spellList.size} spells for class ${className}`);
-
-      const actorLevel = this.actor.system.details.level;
-      const maxSpellLevel = discoveryUtils.calculateMaxSpellLevel(actorLevel, classItem.system.spellcasting);
-
-      const spellItems = await actorSpellUtils.fetchSpellDocuments(spellList, maxSpellLevel);
-      if (!spellItems || !spellItems.length) {
-        log(1, 'No spell items could be loaded');
-        return;
-      }
-      log(3, `Loaded ${spellItems.length} spell items`);
-
-      await this._processAndOrganizeSpells(spellItems, classItem);
-    } catch (error) {
-      log(1, 'Error loading regular spell data:', error);
-    }
-  }
-
-  async _processWizardSpells(allSpellItems, classItem, personalSpellbook) {
-    try {
-      log(3, 'Processing wizard spells for both tabs');
-
-      // Prepare data for both tabs
-      const activeTab = this.tabGroups['spellbook-tabs'];
-
-      // Create an object to hold data for both tabs
-      const tabData = {
-        spellstab: {
-          // Preparation tab - only spells in spellbook and cantrips
-          spellLevels: [],
-          spellPreparation: { current: 0, maximum: 0 }
-        },
-        wizardtab: {
-          // Spellbook tab - all class spells EXCEPT cantrips
-          spellLevels: [],
-          spellPreparation: { current: 0, maximum: 0 }
-        }
-      };
-
-      // Calculate free spells info
-      const totalFreeSpells = this.wizardManager.getTotalFreeSpells();
-      const usedFreeSpells = await this.wizardManager.getUsedFreeSpells();
-      const remainingFreeSpells = Math.max(0, totalFreeSpells - usedFreeSpells);
-      const totalSpells = personalSpellbook.length;
-      log(1, 'Math time:', { totalFree: totalFreeSpells, usedFree: usedFreeSpells, remaining: remainingFreeSpells, totalSpells: totalSpells });
-      // Add to the wizard tab data
-      tabData.wizardtab.wizardTotalSpellbookCount = totalSpells;
-      tabData.wizardtab.wizardFreeSpellbookCount = totalFreeSpells;
-      tabData.wizardtab.wizardRemainingFreeSpells = remainingFreeSpells;
-      tabData.wizardtab.wizardHasFreeSpells = remainingFreeSpells > 0;
-
-      // Identify granted spells on the actor
-      const grantedSpells = this.actor.items
-        .filter((i) => i.type === 'spell' && (i.flags?.dnd5e?.cachedFor || (i.system?.preparation?.mode && ['pact', 'innate', 'atwill'].includes(i.system.preparation.mode))))
-        .map((i) => i.flags?.core?.sourceId || i.uuid)
-        .filter(Boolean);
-
-      log(3, `Found ${grantedSpells.length} granted spells on actor`);
-
-      // For preparation tab - include cantrips, spellbook spells, and granted spells
-      const prepTabSpells = allSpellItems.filter(
-        (spell) =>
-          spell.system.level === 0 || // Include all cantrips
-          personalSpellbook.includes(spell.compendiumUuid) || // Include spells in personal spellbook
-          grantedSpells.includes(spell.compendiumUuid) // Include granted spells
-      );
-
-      // For wizard tab - ONLY include spells from the wizard class spell list
-      const wizardTabSpells = allSpellItems.filter(
-        (spell) =>
-          this._fullWizardSpellList.has(spell.compendiumUuid) && // Only include spells from the official wizard spell list
-          spell.system.level !== 0 // Exclude cantrips
-      );
-
-      // Process spells for both tabs
-      const prepLevels = actorSpellUtils.organizeSpellsByLevel(prepTabSpells, this.actor, this.spellManager);
-
-      // For wizard tab, only use spells from the wizard class list
-      const wizardLevels = actorSpellUtils.organizeSpellsByLevel(wizardTabSpells, null, this.spellManager);
-
-      log(3, `Filtered to ${wizardTabSpells.length} wizard class spells for spellbook tab (${wizardLevels.length} levels)`);
-
-      // Calculate maximum allowed spells and track if at maximum
-      const maxSpellsAllowed = this.wizardManager.getMaxSpellsAllowed();
-      const isAtMaxSpells = personalSpellbook.length >= maxSpellsAllowed;
-
-      // Add to the wizard tab data
-      tabData.wizardtab.wizardMaxSpellbookCount = maxSpellsAllowed;
-      tabData.wizardtab.wizardIsAtMax = isAtMaxSpells;
-
-      // Sort spells within each level
-      const sortBy = this._getFilterState().sortBy || 'level';
-      for (const level of prepLevels) {
-        level.spells = this._sortSpells(level.spells, sortBy);
-      }
-      for (const level of wizardLevels) {
-        // Use wizardLevels here instead of filteredSpellbookLevels
-        level.spells = this._sortSpells(level.spells, sortBy);
-      }
-
-      // Enrich all spells with wizard-specific data
-      for (const level of prepLevels) {
-        for (const spell of level.spells) {
-          spell.isWizardClass = true;
-          spell.inWizardSpellbook = personalSpellbook.includes(spell.compendiumUuid);
-          spell.enrichedIcon = formattingUtils.createSpellIconLink(spell);
-          spell.formattedDetails = formattingUtils.formatSpellDetails(spell);
-        }
-      }
-
-      for (const level of wizardLevels) {
-        // Use wizardLevels here instead of filteredSpellbookLevels
-        for (const spell of level.spells) {
-          spell.isWizardClass = true;
-          spell.inWizardSpellbook = personalSpellbook.includes(spell.compendiumUuid);
-          spell.canAddToSpellbook = !spell.inWizardSpellbook && spell.system.level > 0;
-          spell.enrichedIcon = formattingUtils.createSpellIconLink(spell);
-          spell.formattedDetails = formattingUtils.formatSpellDetails(spell);
-          spell.isAtMaxSpells = isAtMaxSpells;
-        }
-      }
-
-      // Calculate preparation stats
-      const prepStats = this._calculatePreparationStats(prepLevels, classItem);
-
-      // Store data based on active tab
-      tabData.spellstab.spellLevels = prepLevels;
-      tabData.spellstab.spellPreparation = prepStats;
-      tabData.wizardtab.spellLevels = wizardLevels; // Use wizardLevels here instead of filteredSpellbookLevels
-      tabData.wizardtab.spellPreparation = prepStats; // Same stats for both tabs
-
-      // Set the appropriate data based on active tab
-      this.spellLevels = tabData[activeTab].spellLevels;
-      this.spellPreparation = tabData[activeTab].spellPreparation;
-      this.className = classItem.name;
-
-      // Store tab data for quick tab switching without reloading
-      this._tabData = tabData;
-
-      log(3, `Processed ${prepLevels.length} spell levels for preparation tab and ${wizardLevels.length} for spellbook tab`);
-      log(3, `Preparation statistics: ${prepStats.current}/${prepStats.maximum}`);
-    } catch (error) {
-      log(1, 'Error processing wizard spells:', error);
-    }
-  }
-
-  async _loadSpellcastingClass() {
-    try {
-      const classItem = genericUtils.findSpellcastingClass(this.actor);
-      if (!classItem) {
-        return null;
-      }
-
-      log(3, `Found spellcasting class: ${classItem.name}`);
-      return classItem;
-    } catch (error) {
-      log(1, 'Error finding spellcasting class:', error);
-      return null;
-    }
-  }
-
-  async _loadSpellList(classItem) {
-    try {
-      const className = classItem.name.toLowerCase();
-      const classUuid = classItem.uuid;
-
-      log(3, `Loading spell list for ${className}`);
-      // Pass the wizardManager to avoid creating multiple instances
-      const spellUuids = await discoveryUtils.getClassSpellList(className, classUuid, this.actor, this.wizardManager);
-
-      if (!spellUuids || !spellUuids.size) {
-        log(1, 'No spells found in class spell list');
-        return new Set();
-      }
-
-      log(3, `Found ${spellUuids.size} spells for class ${className}`);
-      return spellUuids;
-    } catch (error) {
-      log(1, 'Error loading spell list:', error);
-      return new Set();
-    }
-  }
-
-  async _loadSpellItems(spellUuids, classItem) {
-    try {
-      const actorLevel = this.actor.system.details.level;
-      const maxSpellLevel = discoveryUtils.calculateMaxSpellLevel(actorLevel, classItem.system.spellcasting);
-
-      log(3, `Loading spells up to level ${maxSpellLevel} for ${this.actor.name}`);
-      const spellItems = await actorSpellUtils.fetchSpellDocuments(spellUuids, maxSpellLevel);
-
-      if (!spellItems || !spellItems.length) {
-        return [];
-      }
-
-      log(3, `Loaded ${spellItems.length} spell items`);
-      return spellItems;
-    } catch (error) {
-      log(1, 'Error loading spell items:', error);
-      return [];
-    }
-  }
-
-  async _processAndOrganizeSpells(spellItems, classItem) {
-    try {
-      log(3, 'Processing and organizing spells');
-
-      // Get current active tab
-      const activeTab = this.tabGroups['spellbook-tabs'];
-
-      // For wizards, separate spell list from spellbook
-      let filteredSpellItems = [...spellItems];
-
-      if (this.wizardManager?.isWizard) {
-        const spellbookSpells = this.wizardManager.getSpellbookSpells();
-        log(3, `Wizard has ${spellbookSpells.length} spells in their spellbook`);
-
-        // For the spells tab (preparation), only show spells in the wizard's spellbook plus cantrips
-        if (activeTab === 'spellstab') {
-          // For preparation tab, only show cantrips and spells in their spellbook
-          filteredSpellItems = spellItems.filter(
-            (spell) =>
-              spell.system.level === 0 || // Always show cantrips
-              spellbookSpells.includes(spell.compendiumUuid) // Only show spells in their spellbook
-          );
-          log(3, `Preparation tab: showing ${filteredSpellItems.length} spells in wizard's spellbook`);
-        }
-        // For wizard spellbook tab, we show all available class spells
-        else if (activeTab === 'wizardtab') {
-          // No filtering for the wizard spellbook tab - show all available spells
-          log(3, `Wizard spellbook tab: showing all ${spellItems.length} available class spells`);
-        }
-      }
-
-      // Organize spells by level
-      const spellLevels = actorSpellUtils.organizeSpellsByLevel(filteredSpellItems, this.actor, this.spellManager);
-      log(3, `Organized spells into ${spellLevels.length} levels`);
-
-      // Sort spells within each level
-      const sortBy = this._getFilterState().sortBy || 'level';
-      for (const level of spellLevels) {
-        level.spells = this._sortSpells(level.spells, sortBy);
-      }
-
-      // Add wizard-specific data to each spell
-      if (this.wizardManager?.isWizard) {
-        const spellbookSpells = this.wizardManager.getSpellbookSpells();
-
-        for (const level of spellLevels) {
-          for (const spell of level.spells) {
-            // Mark which spells are in the wizard's spellbook
-            spell.isWizardClass = true;
-            spell.inWizardSpellbook = spellbookSpells.includes(spell.compendiumUuid);
-
-            // Only non-cantrip spells not already in spellbook can be learned
-            spell.canAddToSpellbook = !spell.inWizardSpellbook && spell.system.level > 0;
-
-            // For spells not in spellbook on the preparation tab, disable preparation
-            if (activeTab === 'spellstab' && spell.system.level > 0 && !spell.inWizardSpellbook) {
-              if (spell.preparation) {
-                spell.preparation.disabled = true;
-                spell.preparation.disabledReason = 'SPELLBOOK.Wizard.NotInSpellbook';
-              }
-            }
-          }
-        }
-      }
-
-      // Enrich spell data with icons and details
-      await this._enrichSpellData(spellLevels);
-
-      // Calculate preparation stats
-      const prepStats = this._calculatePreparationStats(spellLevels, classItem);
-
-      // Update state with processed data
-      this.spellLevels = spellLevels;
-      this.className = classItem.name;
-      this.spellPreparation = prepStats;
-
-      log(3, `Preparation statistics: ${prepStats.current}/${prepStats.maximum}`);
-    } catch (error) {
-      log(1, 'Error processing spell data:', error);
-    }
-  }
-
-  async _enrichSpellData(spellLevels) {
-    try {
-      log(3, 'Enriching spell data with icons and details');
-      for (const level of spellLevels) {
-        for (const spell of level.spells) {
-          try {
-            spell.enrichedIcon = formattingUtils.createSpellIconLink(spell);
-            spell.formattedDetails = formattingUtils.formatSpellDetails(spell);
-          } catch (error) {
-            log(1, `Failed to enrich spell: ${spell.name}`, error);
-          }
-        }
-      }
-    } catch (error) {
-      log(1, 'Error enriching spell data:', error);
-    }
-  }
-
-  _calculatePreparationStats(spellLevels, classItem) {
-    try {
-      let preparedCount = 0;
-      let maxPrepared = 0;
-
-      // Calculate maximum prepared spells using the new method
-      if (this.spellManager) {
-        maxPrepared = this.spellManager.getMaxPrepared();
-      } else if (classItem) {
-        // Legacy calculation as fallback
-        const spellcastingAbility = classItem.system.spellcasting?.ability;
-        if (spellcastingAbility) {
-          const abilityMod = this.actor.system.abilities[spellcastingAbility]?.mod || 0;
-          const classLevel = classItem.system.levels || this.actor.system.details.level;
-          maxPrepared = Math.max(1, classLevel + abilityMod);
-        }
-      }
-
-      // Count prepared spells (excluding cantrips)
-      for (const level of spellLevels) {
-        // Skip cantrips (level 0)
-        if (level.level === '0' || level.level === 0) continue;
-
-        for (const spell of level.spells) {
-          if (spell.preparation.prepared && !spell.preparation.alwaysPrepared) {
-            preparedCount++;
-          }
-        }
-      }
-
-      return {
-        current: preparedCount,
-        maximum: maxPrepared
-      };
-    } catch (error) {
-      log(1, 'Error calculating preparation stats:', error);
-      return { current: 0, maximum: 0 };
-    }
-  }
-
-  async _cacheWizardSpellbook() {
-    if (this.wizardManager && this.wizardManager.isWizard) {
-      this.wizardSpellbookCache = await this.wizardManager.getSpellbookSpells();
-      log(3, `Cached ${this.wizardSpellbookCache.length} spells for wizard spellbook`);
-    }
-  }
-  /* -------------------------------------------- */
-  /*  UI Setup & Management                       */
-  /* -------------------------------------------- */
-
-  _setSidebarState() {
-    try {
-      const sidebarCollapsed = game.user.getFlag(MODULE.ID, FLAGS.SIDEBAR_COLLAPSED);
-      if (sidebarCollapsed) {
-        this.element.classList.add('sidebar-collapsed');
-        log(3, 'Setting sidebar to collapsed state');
-      }
-    } catch (error) {
-      log(1, 'Error setting sidebar state:', error);
-    }
-  }
-
-  _disableInputsWhileLoading() {
-    try {
-      const inputs = this.element.querySelectorAll('.spell-filters input, .spell-filters select, .spell-filters button');
-      inputs.forEach((input) => {
-        input.disabled = true;
-      });
-      log(3, 'Disabled inputs during loading');
-    } catch (error) {
-      log(1, 'Error disabling inputs:', error);
-    }
-  }
-
-  _positionFooter() {
-    try {
-      const footer = this.element.querySelector('footer');
-      if (!footer) return;
-
-      const isSidebarCollapsed = this.element.classList.contains('sidebar-collapsed');
-      const sidebarFooterContainer = this.element.querySelector('.sidebar-footer-container');
-      const collapsedFooter = this.element.querySelector('.collapsed-footer');
-
-      if (isSidebarCollapsed && collapsedFooter) {
-        collapsedFooter.appendChild(footer);
-        collapsedFooter.classList.remove('hidden');
-      } else if (sidebarFooterContainer) {
-        sidebarFooterContainer.appendChild(footer);
-        if (collapsedFooter) collapsedFooter.classList.add('hidden');
-      }
-    } catch (error) {
-      log(1, 'Error positioning footer:', error);
-    }
-  }
-
-  _setupFilterListeners() {
-    try {
-      log(3, 'Setting up filter listeners');
-
-      // Use a delegated event handler approach by adding listeners to the filters container
-      const filtersContainer = this.element.querySelector('.spell-filters');
-      if (!filtersContainer) {
-        log(2, 'Filter container not found, unable to set up listeners');
-        return;
-      }
-
-      // Handle 'change' events for select dropdowns and checkboxes
-      filtersContainer.addEventListener('change', (event) => {
-        const target = event.target;
-
-        // Handle different element types
-        if (target.matches('dnd5e-checkbox') || target.matches('select')) {
-          // Apply filters immediately
-          this._applyFilters();
-
-          // Special case for sort-by
-          if (target.name === 'sort-by') {
-            this._applySorting(target.value);
-          }
-        }
-      });
-
-      // Handle 'input' events for text search and number inputs
-      filtersContainer.addEventListener('input', (event) => {
-        const target = event.target;
-        if (target.matches('input[type="text"]')) {
-          // Debounce text search
-          clearTimeout(this._searchTimer);
-          this._searchTimer = setTimeout(() => {
-            this._applyFilters();
-          }, 200);
-        } else if (target.matches('input[type="number"]')) {
-          // Debounce range inputs
-          clearTimeout(this._rangeTimer);
-          this._rangeTimer = setTimeout(() => {
-            this._applyFilters();
-          }, 200);
-        }
-      });
-
-      log(3, 'Delegated filter listeners setup complete');
-    } catch (error) {
-      log(1, 'Error setting up filter listeners:', error);
-    }
-  }
-
-  _setupPreparationListeners() {
-    try {
-      log(3, 'Setting up preparation listeners');
-
-      // Use event delegation for checkbox changes
-      const spellsContainer = this.element.querySelector('.spells-container');
-      if (!spellsContainer) {
-        log(1, 'Spells container not found, unable to set up preparation listeners');
-        return;
-      }
-
-      // Check if we need to handle cantrip rules
-      const isLevelUp = this.spellManager.canBeLeveledUp();
-      log(3, `Level up status during listener setup: ${isLevelUp}`);
-
-      // We need to track original cantrips for level-ups with any rule type
-      if (isLevelUp) {
-        log(3, 'Setting up cantrip tracking for level-up');
-
-        // Store the initial state of cantrips
-        this._cantripTracking = {
-          originalChecked: new Set(),
-          hasUnlearned: false,
-          hasLearned: false,
-          unlearned: null,
-          learned: null
-        };
-
-        // Record initially checked cantrips
-        const cantripItems = spellsContainer.querySelectorAll('.spell-item[data-spell-level="0"]');
-        cantripItems.forEach((item) => {
-          const checkbox = item.querySelector('dnd5e-checkbox[data-uuid]');
-          const spellName = item.querySelector('.spell-name .title')?.textContent || 'unknown';
-
-          if (checkbox && checkbox.checked) {
-            this._cantripTracking.originalChecked.add(checkbox.dataset.uuid);
-            log(3, `Recorded original cantrip: ${spellName}`);
-          }
-        });
-
-        log(3, `Recorded ${this._cantripTracking.originalChecked.size} initial cantrips`);
-      }
-
-      // Use event delegation for all preparation checkboxes
-      spellsContainer.addEventListener('change', async (event) => {
-        const target = event.target;
-
-        // Check if this is a dnd5e-checkbox for spell preparation
-        if (target.matches('dnd5e-checkbox[data-uuid]')) {
-          await this._handlePreparationChange(event);
-        }
-      });
-    } catch (error) {
-      log(1, 'Error setting up preparation listeners:', error);
-    }
-  }
-
-  async _handlePreparationChange(event) {
-    try {
-      // Get checkbox and its data
-      const checkbox = event.target;
-      const uuid = checkbox.dataset.uuid;
-      const spellItem = checkbox.closest('.spell-item');
-      const spellLevel = spellItem?.dataset.spellLevel;
-
-      // Handle cantrip-specific logic
-      if (spellLevel === '0') {
-        await this._handleCantripPreparationChange(event, uuid, spellItem);
-      } else {
-        // Non-cantrip spell - just update UI
-        if (spellItem) {
-          if (checkbox.checked) {
-            spellItem.classList.add('prepared-spell');
-          } else {
-            spellItem.classList.remove('prepared-spell');
-          }
-        }
-      }
-
-      // Update tracking and counts
-      this._updateSpellPreparationTracking();
-      this._updateSpellCounts();
-    } catch (error) {
-      log(1, 'Error handling preparation change:', error);
-    }
-  }
-
-  async _handleCantripPreparationChange(event, uuid, spellItem) {
-    try {
-      // Get checkbox and its state
-      const checkbox = event.target;
-      const isChecked = checkbox.checked;
-      const wasPrepared = checkbox.dataset.wasPrepared === 'true';
-
-      // Get spell name for logging
-      const spellName = spellItem?.querySelector('.spell-name .title')?.textContent || 'unknown';
-
-      log(3, `======== CANTRIP CHANGE ========`);
-      log(3, `Cantrip: ${spellName}, Action: ${isChecked ? 'checking' : 'unchecking'}`);
-
-      // Get context information
-      const isLevelUp = this.spellManager.canBeLeveledUp();
-      const isLongRest = this._isLongRest;
-
-      log(3, `Context: isLevelUp=${isLevelUp}, isLongRest=${isLongRest}`);
-
-      // Get the source spell
-      const sourceSpell = await fromUuid(uuid);
-      if (!sourceSpell) {
-        log(1, `Could not find source spell for UUID: ${uuid}`);
-        return;
-      }
-
-      // Ensure we're using the source UUID consistently
-      const sourceUuid = sourceSpell.flags?.core?.sourceId || sourceSpell.uuid;
-      log(3, `Source UUID: ${sourceUuid}`);
-
-      // Check if the change is allowed
-      const canChange = this.spellManager.canChangeCantripStatus(sourceSpell, isChecked, isLevelUp, isLongRest, this._uiCantripCount);
-
-      // If change is not allowed, revert and notify
-      if (!canChange.allowed) {
-        // Revert the change
-        checkbox.checked = !isChecked;
-
-        // Show notification if provided
-        if (canChange.message) {
-          ui.notifications.warn(game.i18n.localize(canChange.message));
-        }
-
-        // Update counter display
-        this._updateCantripCounter();
-        return;
-      }
-
-      // Track the change for swap management
-      this.spellManager.trackCantripChange(sourceSpell, isChecked, isLevelUp, isLongRest);
-
-      // Track newly checked cantrips for UI
-      if (isChecked && !wasPrepared) {
-        this._newlyCheckedCantrips.add(uuid);
-        log(3, `Added to newly checked cantrips: ${spellName}`);
-      } else if (!isChecked && this._newlyCheckedCantrips.has(uuid)) {
-        this._newlyCheckedCantrips.delete(uuid);
-        log(3, `Removed from newly checked cantrips: ${spellName}`);
-      }
-
-      // Update UI
-      this._updateCantripCounter();
-
-      // Update UI class for spell item
-      if (spellItem) {
-        if (isChecked) {
-          spellItem.classList.add('prepared-spell');
-        } else {
-          spellItem.classList.remove('prepared-spell');
-        }
-      }
-
-      // Update locks for all cantrips
-      this._setupCantripLocks();
-
-      log(3, `Cantrip change completed successfully`);
-    } catch (error) {
-      log(1, `Error handling cantrip preparation change: ${error.message}`);
-    }
-  }
-
-  _updateSpellPreparationTracking() {
-    try {
-      // Find the dnd5e-checkbox elements that are not disabled
-      const preparedCheckboxes = this.element.querySelectorAll('dnd5e-checkbox[data-uuid]:not([disabled])');
-
-      const countDisplay = this.element.querySelector('.spell-prep-tracking');
-      if (!countDisplay) return;
-
-      // Count checked dnd5e-checkbox elements (excluding cantrips)
-      let preparedCount = 0;
-      preparedCheckboxes.forEach((checkbox) => {
-        // Get the spell item
-        const spellItem = checkbox.closest('.spell-item');
-        const spellLevel = spellItem?.dataset.spellLevel;
-
-        // Skip cantrips
-        if (spellLevel === '0') return;
-
-        // DnD5e checkboxes use the 'checked' property or attribute
-        if (checkbox.checked) preparedCount++;
-      });
-
-      // Get maximum from context
-      const maxPrepared = this?.spellPreparation?.maximum || 0;
-
-      // Update counter elements
-      const currentCountEl = countDisplay.querySelector('.current-count');
-      const maxCountEl = countDisplay.querySelector('.max-count');
-
-      if (currentCountEl) currentCountEl.textContent = preparedCount;
-      if (maxCountEl) maxCountEl.textContent = maxPrepared;
-
-      // Add visual indicator when at/over max
-      if (preparedCount >= maxPrepared) {
-        countDisplay.classList.add('at-max');
-      } else {
-        countDisplay.classList.remove('at-max');
-      }
-
-      log(3, `Updated spell preparation tracking: ${preparedCount}/${maxPrepared}`);
-
-      // Only apply limits if we have a valid maximum
-      if (maxPrepared > 0) {
-        if (preparedCount >= maxPrepared) {
-          // Add class to form
-          this.element.classList.add('at-max-spells');
-
-          // Disable unchecked checkboxes (excluding cantrips)
-          const allSpellCheckboxes = this.element.querySelectorAll('dnd5e-checkbox[data-uuid]');
-          allSpellCheckboxes.forEach((checkbox) => {
-            const spellItem = checkbox.closest('.spell-item');
-            const spellLevel = spellItem?.dataset.spellLevel;
-
-            // Don't disable cantrips
-            if (spellLevel === '0') return;
-
-            if (!checkbox.checked) {
-              checkbox.disabled = true;
-              checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Preparation.AtMaximum');
-              spellItem?.classList.add('max-prepared');
-            }
-          });
-        } else {
-          // Remove max spells class
-          this.element.classList.remove('at-max-spells');
-
-          // Re-enable all preparation checkboxes (excluding cantrips and special cases)
-          const allSpellCheckboxes = this.element.querySelectorAll('dnd5e-checkbox[data-uuid]');
-          allSpellCheckboxes.forEach((checkbox) => {
-            const spellItem = checkbox.closest('.spell-item');
-            const spellLevel = spellItem?.dataset.spellLevel;
-
-            // Don't change cantrips
-            if (spellLevel === '0') return;
-
-            // Skip items with special always-prepared or granted tags
-            if (spellItem.querySelector('.tag.always-prepared') || spellItem.querySelector('.tag.granted')) return;
-
-            checkbox.disabled = false;
-            delete checkbox.dataset.tooltip; // Remove the max prepared tooltip
-            spellItem?.classList.remove('max-prepared');
-          });
-        }
-      }
-    } catch (error) {
-      log(1, 'Error updating spell preparation tracking:', error);
-    }
-  }
-
-  _updateSpellCounts() {
-    try {
-      // Get the active tab
-      const activeTab = this.tabGroups['spellbook-tabs'];
-      const activeTabContent = this.element.querySelector(`.tab[data-tab="${activeTab}"]`);
-      if (!activeTabContent) return;
-
-      // Handle wizard tab - remove spell counts completely
-      if (activeTab === 'wizardtab') {
-        const countDisplays = activeTabContent.querySelectorAll('.spell-count');
-        countDisplays.forEach((countDisplay) => {
-          countDisplay.remove();
-        });
-        return;
-      }
-
-      // Process the main spellstab
-      const spellLevels = activeTabContent.querySelectorAll('.spell-level');
-
-      spellLevels.forEach((levelContainer) => {
-        const levelId = levelContainer.dataset.level;
-
-        // Skip cantrips
-        if (levelId === '0') {
-          const countDisplay = levelContainer.querySelector('.spell-count');
-          if (countDisplay) {
-            countDisplay.remove();
-          }
-          return;
-        }
-
-        // Count spells at this level
-        const spellItems = levelContainer.querySelectorAll('.spell-item');
-
-        // Identify countable spells (not always prepared/granted)
-        const countableSpells = [];
-        const preparedSpells = [];
-
-        spellItems.forEach((item) => {
-          const hasAlwaysPrepared = !!item.querySelector('.tag.always-prepared');
-          const hasGranted = !!item.querySelector('.tag.granted');
-          const isPrepared = item.classList.contains('prepared-spell');
-
-          if (!hasAlwaysPrepared && !hasGranted) {
-            countableSpells.push(item);
-
-            if (isPrepared) {
-              preparedSpells.push(item);
-            }
-          }
-        });
-
-        const preparedCount = preparedSpells.length;
-        const totalAvailable = countableSpells.length;
-
-        // Update or create count display
-        const countDisplay = levelContainer.querySelector('.spell-count');
-        if (countDisplay) {
-          if (totalAvailable > 0) {
-            countDisplay.textContent = `(${preparedCount}/${totalAvailable})`;
-          } else {
-            countDisplay.textContent = '';
-          }
-        } else if (totalAvailable > 0) {
-          // Create a new count element
-          const newCount = document.createElement('span');
-          newCount.className = 'spell-count';
-          newCount.setAttribute('aria-label', 'SPELLBOOK.UI.SpellCount');
-          newCount.textContent = `(${preparedCount}/${totalAvailable})`;
-
-          const levelHeading = levelContainer.querySelector('.spell-level-heading');
-          if (levelHeading) {
-            const cantripCounter = levelHeading.querySelector('.cantrip-counter');
-            if (cantripCounter) {
-              levelHeading.insertBefore(newCount, cantripCounter);
-            } else {
-              levelHeading.appendChild(newCount);
-            }
-          }
-        }
-      });
-    } catch (error) {
-      log(1, 'Error updating spell counts:', error);
-    }
-  }
-
-  _applyCollapsedLevels() {
-    try {
-      const collapsedLevels = game.user.getFlag(MODULE.ID, FLAGS.COLLAPSED_LEVELS) || [];
-      log(3, `Applying collapsed states to ${collapsedLevels.length} levels`);
-
-      for (const levelId of collapsedLevels) {
-        const levelContainer = this.element.querySelector(`.spell-level[data-level="${levelId}"]`);
-
-        if (levelContainer) {
-          levelContainer.classList.add('collapsed');
-        }
-      }
-    } catch (error) {
-      log(1, 'Error applying collapsed levels:', error);
-    }
-  }
-
-  _setupCantripUI() {
-    try {
-      log(3, 'Setting up cantrip UI');
-      const cantripLevel = this.element.querySelector('.spell-level[data-level="0"]');
-      if (!cantripLevel) return;
-
-      // Update counters and locks
-      this._updateCantripCounter(cantripLevel);
-      this._setupCantripLocks();
-
-      // Add wizard rules indicator if applicable
-      if (this.wizardManager?.isWizard && this._isLongRest) {
-        // Get cantrip rules
-        const cantripRules = this.spellManager.getSettings().rules;
-
-        // Remove any existing rules info
-        const existingInfo = cantripLevel.querySelector('.wizard-rules-info');
-        if (existingInfo) existingInfo.remove();
-
-        // Create new rules info element
-        const infoElement = document.createElement('div');
-        infoElement.className = 'wizard-rules-info';
-
-        if (cantripRules === CANTRIP_RULES.MODERN_LONG_REST) {
-          infoElement.innerHTML = `<i class="fas fa-info-circle"></i> ${game.i18n.localize('SPELLBOOK.Wizard.ModernCantripRules')}`;
-        } else {
-          infoElement.innerHTML = `<i class="fas fa-info-circle"></i> ${game.i18n.localize('SPELLBOOK.Wizard.LegacyCantripRules')}`;
-        }
-
-        // Add to the heading
-        const levelHeading = cantripLevel.querySelector('.spell-level-heading');
-        if (levelHeading) {
-          levelHeading.appendChild(infoElement);
-        }
-      }
-    } catch (error) {
-      log(1, 'Error setting up cantrip UI:', error);
-    }
-  }
-
-  _updateCantripCounter(cantripLevel) {
-    if (!cantripLevel) {
-      cantripLevel = this.element.querySelector('.spell-level[data-level="0"]');
-    }
-    if (!cantripLevel) return;
-
-    try {
-      const maxCantrips = this.spellManager.getMaxAllowed();
-
-      // Count prepared cantrips using dnd5e-checkbox elements
-      let currentCount = 0;
-      const checkedCantrips = [];
-
-      // Get all cantrip spell items
-      const cantripItems = cantripLevel.querySelectorAll('.spell-item');
-      log(3, `Cantrip counter: found ${cantripItems.length} cantrip items total`);
-
-      // Count manually by examining each item
-      cantripItems.forEach((item) => {
-        const spellName = item.querySelector('.spell-name .title')?.textContent || 'unknown';
-
-        // Always-prepared or granted spells are already counted by the actor
-        if (item.querySelector('.tag.always-prepared') || item.querySelector('.tag.granted')) {
-          log(3, `Skipping always-prepared/granted cantrip: ${spellName}`);
-          return;
-        }
-
-        const checkbox = item.querySelector('dnd5e-checkbox');
-        if (checkbox && checkbox.checked) {
-          currentCount++;
-          checkedCantrips.push(spellName);
-          log(3, `Counted checked cantrip: ${spellName}`);
-        }
-      });
-
-      log(3, `Cantrip counter: total counted=${currentCount}, max=${maxCantrips}`);
-      log(3, `Checked cantrips: ${checkedCantrips.join(', ')}`);
-
-      // Store UI count for validation
-      this._uiCantripCount = currentCount;
-
-      // Update counter display
-      const levelHeading = cantripLevel.querySelector('.spell-level-heading');
-      let counterElem = levelHeading.querySelector('.cantrip-counter');
-
-      if (!counterElem) {
-        counterElem = document.createElement('span');
-        counterElem.className = 'cantrip-counter';
-        const spellCount = levelHeading.querySelector('.spell-count');
-        if (spellCount) {
-          spellCount.after(counterElem);
-        } else {
-          levelHeading.appendChild(counterElem);
-        }
-      }
-
-      counterElem.textContent = `[${currentCount}/${maxCantrips}]`;
-      counterElem.title = game.i18n.localize('SPELLBOOK.Cantrips.CounterTooltip');
-      counterElem.style.display = '';
-
-      if (currentCount >= maxCantrips) {
-        counterElem.classList.add('at-max');
-      } else {
-        counterElem.classList.remove('at-max');
-      }
-
-      // Return the current count and max for convenience
-      return { current: currentCount, max: maxCantrips };
-    } catch (error) {
-      log(1, `Error updating cantrip counter: ${error.message}`);
-      return { current: 0, max: 0 };
-    }
-  }
-
-  _setupCantripLocks() {
-    try {
-      const cantripItems = this.element.querySelectorAll('.spell-item[data-spell-level="0"]');
-      if (!cantripItems.length) return;
-
-      // Get context information
-      const isLevelUp = this.spellManager.canBeLeveledUp();
-      const isLongRest = this._isLongRest;
-      const currentCount = this._uiCantripCount;
-
-      // Use the centralized lock management in SpellManager
-      this.spellManager.lockCantripCheckboxes(cantripItems, isLevelUp, isLongRest, currentCount);
-
-      log(3, 'Cantrip locks updated');
-    } catch (error) {
-      log(1, `Error setting up cantrip locks: ${error.message}`);
-    }
-  }
-
-  _lockAllCantripCheckboxes() {
-    try {
-      log(3, 'Locking all cantrip checkboxes after swap completion');
-      const cantripItems = this.element.querySelectorAll('.spell-item[data-spell-level="0"]');
-
-      for (const item of cantripItems) {
-        const checkbox = item.querySelector('dnd5e-checkbox');
-        if (!checkbox || checkbox.hasAttribute('data-always-disabled')) continue;
-
-        // Lock everything by setting disabled property and tooltip
-        checkbox.disabled = true;
-        checkbox.dataset.tooltip = game.i18n.localize('SPELLBOOK.Cantrips.SwapComplete');
-        item.classList.add('cantrip-locked');
-
-        // Remove existing lock icons if present
-        const lockIcon = item.querySelector('.cantrip-lock-icon');
-        if (lockIcon) lockIcon.remove();
-      }
-    } catch (error) {
-      log(1, 'Error locking cantrip checkboxes:', error);
-    }
-  }
-
-  setLongRestContext(isLongRest) {
-    // Convert to boolean and log
-    const wasLongRest = this._isLongRest;
-    this._isLongRest = !!isLongRest;
-
-    log(3, `Long rest context changed from ${wasLongRest} to ${this._isLongRest}`);
-
-    // Only update if there's an actual change
-    if (wasLongRest !== this._isLongRest) {
-      if (this._isLongRest) {
-        // Set the flag when activating long rest mode
-        log(3, `Setting long rest flag on actor ${this.actor.name}`);
-        this.actor
-          .setFlag(MODULE.ID, FLAGS.WIZARD_LONG_REST_TRACKING, true)
-          .then(() => {
-            log(3, `Long rest flag successfully set for ${this.actor.name}`);
-          })
-          .catch((err) => {
-            log(1, `Error setting long rest flag: ${err.message}`);
-          });
-      }
-
-      this.render(false); // Re-render with new context
-    }
-  }
-
-  /* -------------------------------------------- */
-  /*  Filter & Sort Methods                       */
-  /* -------------------------------------------- */
-
-  _getFilterState() {
-    if (!this.element) {
-      return filterUtils.getDefaultFilterState();
-    }
-
-    return {
-      name: this.element.querySelector('[name="filter-name"]')?.value || '',
-      level: this.element.querySelector('[name="filter-level"]')?.value || '',
-      school: this.element.querySelector('[name="filter-school"]')?.value || '',
-      castingTime: this.element.querySelector('[name="filter-castingTime"]')?.value || '',
-      minRange: this.element.querySelector('[name="filter-min-range"]')?.value || '',
-      maxRange: this.element.querySelector('[name="filter-max-range"]')?.value || '',
-      damageType: this.element.querySelector('[name="filter-damageType"]')?.value || '',
-      condition: this.element.querySelector('[name="filter-condition"]')?.value || '',
-      requiresSave: this.element.querySelector('[name="filter-requiresSave"]')?.value || '',
-      prepared: this.element.querySelector('[name="filter-prepared"]')?.checked || false,
-      ritual: this.element.querySelector('[name="filter-ritual"]')?.checked || false,
-      concentration: this.element.querySelector('[name="filter-concentration"]')?.value || '',
-      sortBy: this.element.querySelector('[name="sort-by"]')?.value || 'level'
-    };
-  }
-
   _prepareFilters() {
     try {
-      log(3, 'Preparing filters for display');
-
-      // Get the filter configuration
       let filterConfig = game.settings.get(MODULE.ID, SETTINGS.FILTER_CONFIGURATION);
       if (!Array.isArray(filterConfig) || !filterConfig.length) {
         filterConfig = DEFAULT_FILTER_CONFIG;
       }
 
-      // Get enabled filters in the right order
       const sortedFilters = filterConfig.filter((f) => f.enabled).sort((a, b) => a.order - b.order);
+      const filterState = this.filterHelper.getFilterState();
 
-      const filterState = this._getFilterState();
-
-      // Process each filter
       return sortedFilters.map((filter) => {
         const result = {
           id: filter.id,
@@ -1628,10 +374,8 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
           label: game.i18n.localize(filter.label)
         };
 
-        // Create elements based on filter type
         let element;
 
-        // Add type-specific properties
         switch (filter.type) {
           case 'search':
             element = formElements.createTextInput({
@@ -1643,7 +387,7 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
             break;
 
           case 'dropdown':
-            const options = filterUtils.getOptionsForFilter(filter.id, filterState, this.spellLevels);
+            const options = this._getFilterOptions(filter.id, filterState);
             element = formElements.createSelect({
               name: `filter-${filter.id}`,
               options: options,
@@ -1661,46 +405,12 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
             break;
 
           case 'range':
-            // For range filters, create a container with two inputs
-            const container = document.createElement('div');
-            container.className = 'range-inputs';
-            container.setAttribute('role', 'group');
-            container.setAttribute('aria-labelledby', `${filter.id}-label`);
-
-            // Min input
-            const minInput = formElements.createNumberInput({
-              name: `filter-min-range`,
-              value: filterState.minRange || '',
-              placeholder: game.i18n.localize('SPELLBOOK.Filters.RangeMin'),
-              ariaLabel: game.i18n.localize('SPELLBOOK.Filters.RangeMinLabel')
-            });
-
-            // Separator
-            const separator = document.createElement('div');
-            separator.className = 'range-separator';
-            separator.setAttribute('aria-hidden', 'true');
-            separator.innerHTML = '<dnd5e-icon src="systems/dnd5e/icons/svg/range-connector.svg"></dnd5e-icon>';
-
-            // Max input
-            const maxInput = formElements.createNumberInput({
-              name: `filter-max-range`,
-              value: filterState.maxRange || '',
-              placeholder: game.i18n.localize('SPELLBOOK.Filters.RangeMax'),
-              ariaLabel: game.i18n.localize('SPELLBOOK.Filters.RangeMaxLabel')
-            });
-
-            container.appendChild(minInput);
-            container.appendChild(separator);
-            container.appendChild(maxInput);
-
-            element = container;
+            element = this._createRangeFilterElement(filter.id, filterState);
             result.unit = game.settings.get(MODULE.ID, SETTINGS.DISTANCE_UNIT);
             break;
         }
 
-        // Convert element to HTML string
         result.elementHtml = formElements.elementToHtml(element);
-
         return result;
       });
     } catch (error) {
@@ -1709,272 +419,110 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
     }
   }
 
-  _applyFilters() {
-    try {
-      const filters = this._getFilterState();
-      log(3, 'Applying filters to spell list');
-
-      const spellItems = this.element.querySelectorAll('.spell-item');
-      let visibleCount = 0;
-
-      // Track visible spells per level
-      const levelVisibilityMap = new Map();
-
-      for (const item of spellItems) {
-        // Get basic spell metadata
-        const nameEl = item.querySelector('.spell-name');
-        const name = nameEl?.textContent.toLowerCase() || '';
-        const isPrepared = item.classList.contains('prepared-spell');
-        const level = item.dataset.spellLevel || '';
-        const school = item.dataset.spellSchool || '';
-
-        // Get filter data from dataset
-        const castingTimeType = item.dataset.castingTimeType || '';
-        const castingTimeValue = item.dataset.castingTimeValue || '';
-        const rangeUnits = item.dataset.rangeUnits || '';
-        const damageTypes = (item.dataset.damageTypes || '').split(',');
-        const isRitual = item.dataset.ritual === 'true';
-        const isConcentration = item.dataset.concentration === 'true';
-        const requiresSave = item.dataset.requiresSave === 'true';
-        const conditions = (item.dataset.conditions || '').split(',');
-
-        // Special preparation statuses
-        const isGranted = !!item.querySelector('.tag.granted');
-        const isAlwaysPrepared = !!item.querySelector('.tag.always-prepared');
-        const isCountable = !isGranted && !isAlwaysPrepared;
-
-        // Apply each filter as a separate condition
-        const visible = this._checkSpellVisibility(filters, {
-          name,
-          isPrepared,
-          level,
-          school,
-          castingTimeType,
-          castingTimeValue,
-          rangeUnits,
-          rangeValue: item.dataset.rangeValue || '0',
-          damageTypes,
-          isRitual,
-          isConcentration,
-          requiresSave,
-          conditions
-        });
-
-        // Update visibility
-        item.style.display = visible ? '' : 'none';
-
-        if (visible) {
-          visibleCount++;
-
-          // Update level visibility tracker
-          if (!levelVisibilityMap.has(level)) {
-            levelVisibilityMap.set(level, {
-              visible: 0,
-              prepared: 0,
-              countable: 0,
-              countablePrepared: 0
-            });
-          }
-
-          const levelStats = levelVisibilityMap.get(level);
-          levelStats.visible++;
-
-          if (isCountable) {
-            levelStats.countable++;
-            if (isPrepared) levelStats.countablePrepared++;
-          }
-
-          if (isPrepared) levelStats.prepared++;
-        }
-      }
-
-      // Show/hide no results message
-      const noResults = this.element.querySelector('.no-filter-results');
-      if (noResults) {
-        noResults.style.display = visibleCount > 0 ? 'none' : 'block';
-      }
-
-      // Update level container visibility and counts
-      this._updateLevelContainers(levelVisibilityMap);
-
-      log(3, `Filter applied: ${visibleCount} spells visible`);
-    } catch (error) {
-      log(1, 'Error applying filters:', error);
-    }
+  _getFilterOptions(filterId, filterState) {
+    return filterUtils.getOptionsForFilter(filterId, filterState, this.spellLevels);
   }
 
-  _checkSpellVisibility(filters, spell) {
-    // Text search
-    if (filters.name && !spell.name.includes(filters.name.toLowerCase())) {
-      return false;
-    }
+  _createRangeFilterElement(filterId, filterState) {
+    const container = document.createElement('div');
+    container.className = 'range-inputs';
+    container.setAttribute('role', 'group');
+    container.setAttribute('aria-labelledby', `${filterId}-label`);
 
-    // Level filter
-    if (filters.level && spell.level !== filters.level) {
-      return false;
-    }
-
-    // School filter
-    if (filters.school && spell.school !== filters.school) {
-      return false;
-    }
-
-    // Casting Time filter
-    if (filters.castingTime) {
-      const [filterType, filterValue] = filters.castingTime.split(':');
-      const itemType = spell.castingTimeType;
-      const itemValue = spell.castingTimeValue === '' || spell.castingTimeValue === null ? '1' : spell.castingTimeValue;
-
-      if (itemType !== filterType || itemValue !== filterValue) {
-        return false;
-      }
-    }
-
-    // Range filter
-    if ((filters.minRange || filters.maxRange) && spell.rangeUnits) {
-      const rangeValue = parseInt(spell.rangeValue, 10);
-      const convertedRange = filterUtils.convertRangeToStandardUnit(spell.rangeUnits, rangeValue);
-
-      const minRange = filters.minRange ? parseInt(filters.minRange, 10) : 0;
-      const maxRange = filters.maxRange ? parseInt(filters.maxRange, 10) : Infinity;
-
-      if (convertedRange < minRange || convertedRange > maxRange) {
-        return false;
-      }
-    }
-
-    // Damage Type filter
-    if (filters.damageType && !spell.damageTypes.includes(filters.damageType)) {
-      return false;
-    }
-
-    // Condition filter
-    if (filters.condition && !spell.conditions.includes(filters.condition)) {
-      return false;
-    }
-
-    // Requires Save filter
-    if (filters.requiresSave) {
-      if (filters.requiresSave === 'true' && !spell.requiresSave) {
-        return false;
-      } else if (filters.requiresSave === 'false' && spell.requiresSave) {
-        return false;
-      }
-    }
-
-    // Prepared only
-    if (filters.prepared && !spell.isPrepared) {
-      return false;
-    }
-
-    // Ritual only
-    if (filters.ritual && !spell.isRitual) {
-      return false;
-    }
-
-    // Concentration filter
-    if (filters.concentration) {
-      if (filters.concentration === 'true' && !spell.isConcentration) {
-        return false;
-      } else if (filters.concentration === 'false' && spell.isConcentration) {
-        return false;
-      }
-    }
-
-    return true;
-  }
-
-  _sortSpells(spells, sortBy) {
-    return [...spells].sort((a, b) => {
-      switch (sortBy) {
-        case 'name':
-          return a.name.localeCompare(b.name);
-
-        case 'school':
-          const schoolA = a.system.school || '';
-          const schoolB = b.system.school || '';
-          return schoolA.localeCompare(schoolB) || a.name.localeCompare(b.name);
-
-        case 'prepared':
-          // Sort prepared spells first, then by name
-          const prepA = a.preparation.prepared ? 0 : 1;
-          const prepB = b.preparation.prepared ? 0 : 1;
-          return prepA - prepB || a.name.localeCompare(b.name);
-
-        case 'level':
-        default:
-          return a.name.localeCompare(b.name);
-      }
+    const minInput = formElements.createNumberInput({
+      name: `filter-min-range`,
+      value: filterState.minRange || '',
+      placeholder: game.i18n.localize('SPELLBOOK.Filters.RangeMin'),
+      ariaLabel: game.i18n.localize('SPELLBOOK.Filters.RangeMinLabel')
     });
+
+    const separator = document.createElement('div');
+    separator.className = 'range-separator';
+    separator.setAttribute('aria-hidden', 'true');
+    separator.innerHTML = '<dnd5e-icon src="systems/dnd5e/icons/svg/range-connector.svg"></dnd5e-icon>';
+
+    const maxInput = formElements.createNumberInput({
+      name: `filter-max-range`,
+      value: filterState.maxRange || '',
+      placeholder: game.i18n.localize('SPELLBOOK.Filters.RangeMax'),
+      ariaLabel: game.i18n.localize('SPELLBOOK.Filters.RangeMaxLabel')
+    });
+
+    container.appendChild(minInput);
+    container.appendChild(separator);
+    container.appendChild(maxInput);
+
+    return container;
   }
 
-  _applySorting(sortBy) {
+  _applyFilters() {
+    this.filterHelper.applyFilters();
+  }
+
+  async _handlePreparationChange(event) {
     try {
-      log(3, `Applying sorting: ${sortBy}`);
-      const levelContainers = this.element.querySelectorAll('.spell-level');
+      const checkbox = event.target;
+      const uuid = checkbox.dataset.uuid;
+      const spellItem = checkbox.closest('.spell-item');
+      const spellLevel = spellItem?.dataset.spellLevel;
 
-      for (const levelContainer of levelContainers) {
-        const list = levelContainer.querySelector('.spell-list');
-        if (!list) continue;
-
-        const items = Array.from(list.children);
-
-        items.sort((a, b) => {
-          switch (sortBy) {
-            case 'name':
-              return a.querySelector('.spell-name').textContent.localeCompare(b.querySelector('.spell-name').textContent);
-
-            case 'school':
-              const schoolA = a.dataset.spellSchool || '';
-              const schoolB = b.dataset.spellSchool || '';
-              return schoolA.localeCompare(schoolB) || a.querySelector('.spell-name').textContent.localeCompare(b.querySelector('.spell-name').textContent);
-
-            case 'prepared':
-              const aPrepared = a.classList.contains('prepared-spell') ? 0 : 1;
-              const bPrepared = b.classList.contains('prepared-spell') ? 0 : 1;
-              return aPrepared - bPrepared || a.querySelector('.spell-name').textContent.localeCompare(b.querySelector('.spell-name').textContent);
-
-            default:
-              return 0; // Keep current order
-          }
-        });
-
-        // Re-append sorted items
-        for (const item of items) {
-          list.appendChild(item);
-        }
+      if (spellLevel === '0') {
+        await this._handleCantripPreparationChange(event, uuid, spellItem);
+      } else if (spellItem) {
+        spellItem.classList.toggle('prepared-spell', checkbox.checked);
       }
+
+      this.ui.updateSpellPreparationTracking();
+      this.ui.updateSpellCounts();
     } catch (error) {
-      log(1, 'Error applying sorting:', error);
+      log(1, 'Error handling preparation change:', error);
     }
   }
 
-  _updateLevelContainers(levelVisibilityMap) {
+  async _handleCantripPreparationChange(event, uuid, spellItem) {
     try {
-      const levelContainers = this.element.querySelectorAll('.spell-level');
+      const checkbox = event.target;
+      const isChecked = checkbox.checked;
+      const wasPrepared = checkbox.dataset.wasPrepared === 'true';
+      const spellName = spellItem?.querySelector('.spell-name .title')?.textContent || 'unknown';
 
-      for (const container of levelContainers) {
-        const levelId = container.dataset.level;
-        const levelStats = levelVisibilityMap.get(levelId) || {
-          visible: 0,
-          prepared: 0,
-          countable: 0,
-          countablePrepared: 0
-        };
+      const isLevelUp = this.spellManager.canBeLeveledUp();
+      const isLongRest = this._isLongRest;
 
-        // Update visibility
-        container.style.display = levelStats.visible > 0 ? '' : 'none';
-
-        // Update count display
-        const countDisplay = container.querySelector('.spell-count');
-        if (countDisplay && levelStats.countable > 0) {
-          countDisplay.textContent = `(${levelStats.countablePrepared}/${levelStats.countable})`;
-        } else if (countDisplay) {
-          countDisplay.textContent = '';
-        }
+      const sourceSpell = await fromUuid(uuid);
+      if (!sourceSpell) {
+        log(1, `Could not find source spell for UUID: ${uuid}`);
+        return;
       }
+
+      const canChange = this.spellManager.canChangeCantripStatus(sourceSpell, isChecked, isLevelUp, isLongRest, this._uiCantripCount);
+
+      if (!canChange.allowed) {
+        checkbox.checked = !isChecked;
+        if (canChange.message) {
+          ui.notifications.warn(game.i18n.localize(canChange.message));
+        }
+        this.ui.updateCantripCounter();
+        return;
+      }
+
+      this.spellManager.trackCantripChange(sourceSpell, isChecked, isLevelUp, isLongRest);
+
+      if (isChecked && !wasPrepared) {
+        this._newlyCheckedCantrips.add(uuid);
+      } else if (!isChecked && this._newlyCheckedCantrips.has(uuid)) {
+        this._newlyCheckedCantrips.delete(uuid);
+      }
+
+      this.ui.updateCantripCounter();
+
+      if (spellItem) {
+        spellItem.classList.toggle('prepared-spell', isChecked);
+      }
+
+      this.ui.setupCantripLocks();
     } catch (error) {
-      log(1, 'Error updating level containers:', error);
+      log(1, 'Error handling cantrip preparation change:', error);
     }
   }
 
@@ -1984,20 +532,15 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
 
   static toggleSidebar(event, _form) {
     try {
-      log(3, 'Toggling sidebar');
       const isCollapsing = !this.element.classList.contains('sidebar-collapsed');
       this.element.classList.toggle('sidebar-collapsed');
 
-      // Rotate the caret icon
       const caretIcon = event.currentTarget.querySelector('i');
       if (caretIcon) {
         caretIcon.style.transform = isCollapsing ? 'rotate(180deg)' : 'rotate(0)';
       }
 
-      // Reposition footer
-      this._positionFooter();
-
-      // Store user preference
+      this.ui.positionFooter();
       game.user.setFlag(MODULE.ID, FLAGS.SIDEBAR_COLLAPSED, isCollapsing);
     } catch (error) {
       log(1, 'Error toggling sidebar:', error);
@@ -2005,19 +548,13 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
   }
 
   static filterSpells(_event, _form) {
-    try {
-      log(3, 'Filtering spells');
-      this._applyFilters();
-    } catch (error) {
-      log(1, 'Error filtering spells:', error);
-    }
+    this._applyFilters();
   }
 
   static sortSpells(event, _form) {
     try {
-      log(3, 'Sorting spells');
       const sortBy = event.target.value;
-      this._applySorting(sortBy);
+      this.filterHelper.applySorting(sortBy);
     } catch (error) {
       log(1, 'Error sorting spells:', error);
     }
@@ -2025,97 +562,59 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
 
   static handleReset(event, form) {
     try {
-      log(3, 'Handling form reset');
-
-      // Check if shift key is pressed for alternative reset
       const isShiftReset = event.shiftKey;
 
       if (isShiftReset) {
-        // Alternative reset: uncheck all boxes
-        log(3, 'Performing alternative reset (uncheck all)');
-
-        // Uncheck all non-disabled preparation checkboxes
         const checkboxes = this.element.querySelectorAll('dnd5e-checkbox[data-uuid]:not([disabled])');
-        checkboxes.forEach((checkbox) => {
-          checkbox.checked = false;
-        });
+        checkboxes.forEach((checkbox) => (checkbox.checked = false));
 
-        // Reset filters to default state
         const filters = this.element.querySelectorAll('.spell-filters input, .spell-filters select');
         filters.forEach((filter) => {
-          if (filter.type === 'checkbox') {
-            filter.checked = false;
-          } else if (filter.type === 'text' || filter.type === 'number') {
-            filter.value = '';
-          } else if (filter.tagName === 'SELECT') {
-            filter.selectedIndex = 0;
-          }
+          if (filter.type === 'checkbox') filter.checked = false;
+          else if (filter.type === 'text' || filter.type === 'number') filter.value = '';
+          else if (filter.tagName === 'SELECT') filter.selectedIndex = 0;
         });
 
-        // Update UI classes
         const spellItems = this.element.querySelectorAll('.spell-item');
         spellItems.forEach((item) => {
-          // Only remove the class from non-disabled items
           const checkbox = item.querySelector('dnd5e-checkbox');
-          if (checkbox && !checkbox.disabled) {
-            item.classList.remove('prepared-spell');
-          }
+          if (checkbox && !checkbox.disabled) item.classList.remove('prepared-spell');
         });
 
-        // Uncollapse all spell levels
         const collapsedLevels = this.element.querySelectorAll('.spell-level.collapsed');
         collapsedLevels.forEach((level) => {
           level.classList.remove('collapsed');
-
-          // Update the aria-expanded attribute
           const heading = level.querySelector('.spell-level-heading');
-          if (heading) {
-            heading.setAttribute('aria-expanded', 'true');
-          }
+          if (heading) heading.setAttribute('aria-expanded', 'true');
         });
 
-        // Clear the collapsed levels in user flags
         game.user.setFlag(MODULE.ID, FLAGS.COLLAPSED_LEVELS, []);
 
-        // Reapply filters and update tracking
         this._applyFilters();
-        this._updateSpellPreparationTracking();
-        this._updateCantripCounter();
+        this.ui.updateSpellPreparationTracking();
+        this.ui.updateCantripCounter();
 
-        // Prevent default reset behavior
         event.preventDefault();
       } else {
-        // Original reset behavior
-        // Give the browser time to reset form elements
         setTimeout(() => {
-          // Update spell items to match checkbox state
           const spellItems = this.element.querySelectorAll('.spell-item');
           spellItems.forEach((item) => {
             const checkbox = item.querySelector('dnd5e-checkbox');
-            if (checkbox && !checkbox.checked) {
-              item.classList.remove('prepared-spell');
-            }
+            if (checkbox && !checkbox.checked) item.classList.remove('prepared-spell');
           });
 
-          // Uncollapse all spell levels
           const collapsedLevels = this.element.querySelectorAll('.spell-level.collapsed');
           collapsedLevels.forEach((level) => {
             level.classList.remove('collapsed');
-
-            // Update the aria-expanded attribute
             const heading = level.querySelector('.spell-level-heading');
-            if (heading) {
-              heading.setAttribute('aria-expanded', 'true');
-            }
+            if (heading) heading.setAttribute('aria-expanded', 'true');
           });
 
-          // Clear the collapsed levels in user flags
           game.user.setFlag(MODULE.ID, FLAGS.COLLAPSED_LEVELS, []);
 
-          // Reapply filters and update tracking
           this._applyFilters();
-          this._updateSpellPreparationTracking();
-          this._updateCantripCounter();
+          this.ui.updateSpellPreparationTracking();
+          this.ui.updateCantripCounter();
         }, 0);
       }
     } catch (error) {
@@ -2126,17 +625,11 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
   static toggleSpellLevel(_event, form) {
     try {
       const levelContainer = form.parentElement;
-      if (!levelContainer || !levelContainer.classList.contains('spell-level')) {
-        return;
-      }
+      if (!levelContainer || !levelContainer.classList.contains('spell-level')) return;
 
       const levelId = levelContainer.dataset.level;
-      log(3, `Toggling spell level: ${levelId}`);
-
-      // Toggle collapsed state
       levelContainer.classList.toggle('collapsed');
 
-      // Save state to user flags
       const collapsedLevels = game.user.getFlag(MODULE.ID, FLAGS.COLLAPSED_LEVELS) || [];
       const isCollapsed = levelContainer.classList.contains('collapsed');
 
@@ -2154,7 +647,6 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
 
   static configureFilters(_event, _form) {
     try {
-      log(3, 'Opening filter configuration');
       const filterConfig = new PlayerFilterConfiguration(this);
       filterConfig.render(true);
     } catch (error) {
@@ -2164,7 +656,6 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
 
   static configureCantripSettings(_event, _form) {
     try {
-      log(3, 'Opening cantrip settings configuration');
       const dialog = new SpellbookSettingsDialog(this.actor);
       dialog.render(true);
     } catch (error) {
@@ -2177,22 +668,18 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
       const spellUuid = event.target.dataset.uuid;
       if (!spellUuid) return;
 
-      // Store UI state
       const collapsedLevels = Array.from(this.element.querySelectorAll('.spell-level.collapsed')).map((el) => el.dataset.level);
       const activeTab = this.tabGroups['spellbook-tabs'];
 
-      // Load spell
       const spell = await fromUuid(spellUuid);
       if (!spell) {
         ui.notifications.error(`Could not find spell with UUID: ${spellUuid}`);
         return;
       }
 
-      // Get cost information
       const costInfo = await this.wizardManager.getCopyingCostWithFree(spell);
       const time = this.wizardManager.getCopyingTime(spell);
 
-      // Show confirmation dialog
       const dialog = new WizardSpellCopyDialog(spell, this.wizardManager, costInfo, time);
       const result = await dialog.getResult();
 
@@ -2202,22 +689,12 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
         if (success) {
           ui.notifications.info(`Learned ${spell.name} and added it to your spellbook.`);
 
-          // Update local cache
-          if (this.wizardSpellbookCache) {
-            this.wizardSpellbookCache.push(spellUuid);
+          if (this._stateManager.wizardSpellbookCache) {
+            this._stateManager.wizardSpellbookCache.push(spellUuid);
           }
 
-          // Update wizard counts in tabData immediately
-          if (this._tabData && this._tabData.wizardtab) {
-            this._tabData.wizardtab.wizardTotalSpellbookCount = (this._tabData.wizardtab.wizardTotalSpellbookCount || 0) + 1;
+          this._updateWizardTabDataAfterSpellLearning(costInfo.isFree);
 
-            if (costInfo.isFree) {
-              this._tabData.wizardtab.wizardRemainingFreeSpells = Math.max(0, (this._tabData.wizardtab.wizardRemainingFreeSpells || 0) - 1);
-              this._tabData.wizardtab.wizardHasFreeSpells = this._tabData.wizardtab.wizardRemainingFreeSpells > 0;
-            }
-          }
-
-          // Update UI: Mark spell as in spellbook
           const spellItem = this.element.querySelector(`.spell-item[data-spell-uuid="${spellUuid}"]`);
           if (spellItem) {
             const buttonContainer = spellItem.querySelector('.wizard-spell-status');
@@ -2227,20 +704,14 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
             spellItem.classList.add('in-wizard-spellbook', 'prepared-spell');
           }
 
-          // Flag for prep tab updating
           this._spellsTabNeedsReload = true;
-
-          // Full re-render (needed for template updates)
           this.render(false);
 
-          // Restore UI state after render
           setTimeout(() => {
-            // Restore active tab
             if (activeTab && this.tabGroups['spellbook-tabs'] !== activeTab) {
               this.changeTab(activeTab, 'spellbook-tabs');
             }
 
-            // Restore collapsed levels
             collapsedLevels.forEach((levelId) => {
               const levelEl = this.element.querySelector(`.spell-level[data-level="${levelId}"]`);
               if (levelEl) {
@@ -2255,21 +726,27 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
         }
       }
     } catch (error) {
-      log(1, `Error learning spell: ${error.message}`);
+      log(1, 'Error learning spell:', error);
       ui.notifications.error('Failed to learn spell');
+    }
+  }
+
+  _updateWizardTabDataAfterSpellLearning(isFree) {
+    if (this._stateManager.tabData && this._stateManager.tabData.wizardtab) {
+      this._stateManager.tabData.wizardtab.wizardTotalSpellbookCount = (this._stateManager.tabData.wizardtab.wizardTotalSpellbookCount || 0) + 1;
+
+      if (isFree) {
+        this._stateManager.tabData.wizardtab.wizardRemainingFreeSpells = Math.max(0, (this._stateManager.tabData.wizardtab.wizardRemainingFreeSpells || 0) - 1);
+        this._stateManager.tabData.wizardtab.wizardHasFreeSpells = this._stateManager.tabData.wizardtab.wizardRemainingFreeSpells > 0;
+      }
     }
   }
 
   static async formHandler(_event, form, formData) {
     try {
-      log(3, 'Processing form submission');
       const actor = this.actor;
-      if (!actor) {
-        log(3, 'No actor found');
-        return null;
-      }
+      if (!actor) return null;
 
-      // Create a comprehensive spellData object for ALL checkboxes
       const spellData = {};
       const checkboxes = form.querySelectorAll('dnd5e-checkbox[data-uuid]');
 
@@ -2278,57 +755,36 @@ export class PlayerSpellBook extends HandlebarsApplicationMixin(ApplicationV2) {
         const name = checkbox.dataset.name;
         const wasPrepared = checkbox.dataset.wasPrepared === 'true';
         const isPrepared = checkbox.checked;
-        const isRitual = checkbox.dataset.ritual === 'true'; // Add ritual status
+        const isRitual = checkbox.dataset.ritual === 'true';
 
-        // Check if this is an always-prepared spell
         const spellItem = checkbox.closest('.spell-item');
         const isAlwaysPreparedElement = spellItem && (spellItem.querySelector('.tag.always-prepared') || spellItem.querySelector('.tag.granted'));
 
-        // If it's an always-prepared spell or granted spell, don't include it in the updates
-        if (isAlwaysPreparedElement) {
-          log(3, `Skipping always-prepared or granted spell: ${name}`);
-          continue;
-        }
+        if (isAlwaysPreparedElement) continue;
 
-        log(3, `Spell ${name} (${uuid}): was ${wasPrepared ? 'prepared' : 'not prepared'}, now ${isPrepared ? 'prepared' : 'not prepared'}, ritual: ${isRitual}`);
-
-        // Add the spell to our update data - include ritual status
         spellData[uuid] = {
           name,
           wasPrepared,
           isPrepared,
           isAlwaysPrepared: false,
-          isRitual // Add ritual status
+          isRitual
         };
       }
 
-      // Use SpellManager to save prepared spells
       await this.spellManager.saveActorPreparedSpells(spellData);
 
-      // Check if we need to finalize a cantrip level-up
       if (this.spellManager.canBeLeveledUp()) {
         await this.spellManager.completeCantripsLevelUp();
-        log(3, 'Finalized cantrip level-up selection');
       }
 
-      // If we were in long rest mode, clear all tracking data
       if (this._isLongRest) {
-        log(3, 'Clearing long rest cantrip swap flags');
-
-        // Reset all swap tracking data
         await this.spellManager.resetSwapTracking();
-
-        // Set the long rest tracking flag to FALSE instead of removing it
         await actor.setFlag(MODULE.ID, FLAGS.WIZARD_LONG_REST_TRACKING, false);
-        log(3, 'Set wizardLongRestTracking flag to FALSE');
-
         this._isLongRest = false;
       }
 
-      // Clear tracking of newly checked cantrips
       this._newlyCheckedCantrips.clear();
 
-      // Re-render character sheet if open
       if (actor.sheet.rendered) {
         actor.sheet.render(true);
       }
